Move battery function to battery module
