+++ app/src/log_module.c
@@ -8,7 +8,6 @@
  * Author: Ashwin Whitchurch, Protocentral Electronics
  * Contact: ashwin@protocentral.com
  *
- *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
